// src/types/data/project/org-project-type.ts
export interface GetSingleOrgProjectItemShape {
  id: number;
  motherId: number;
  projectName: string;
  projectCode: string;
}

// table
export interface GetSingleOrgProjectTableItemShape {
  id: 1;
  projectCode: "1";
  projectName: "بدون پروژه درآمد";
  dateFrom: "";
  dateEnd: "";
  areaArray: ".1.2.3.4.5.6.7.8.9.11.12.13.14.15.16.17.18.19.20.21.22.23.24.25.26.27.28.29";
  projectScaleId: 3;
  projectScaleName: "منطقه ای";
  dateFromShamsi: "string";
  dateEndShamsi: "string";
}